refactor(connection): extract useOutgoingConnectionRequest hook
ConnectToPeer no longer owns the outgoing-request business logic. The new
useOutgoingConnectionRequest hook holds the waiting state derived from the
snapshot, the minimum-wait hold, the transient rejection toast, and the
connect/cancel actions. ConnectToPeer is reduced to the token input, the
waiting UI, and form handling.
